+++ core/allotment.go
@@ -6,13 +6,14 @@ import (
 	"math/big"
 )
 
+type Allotment []big.Rat
+
 func NewAllotment(portions []Portion) (*Allotment, error) {
 	n := len(portions)
 	total := big.NewRat(0, 1)
 	var remaining_idx *int
 	allotment := make([]big.Rat, n)
 	for i := 0; i < n; i++ {
-		fmt.Println(portions[i])
 		if portions[i].Remaining {
 			if remaining_idx != nil {
 				return nil, errors.New("two uses of `remaining` in the same allotment")
@@ -39,4 +40,12 @@ func NewAllotment(portions []Portion) (*Allotment, error) {
 	}
 	result := Allotment(allotment)
 	return &result, nil
+}
+
+func (a Allotment) String() string {
+	out := "{\n"
+	for _, ratio := range a {
+		out += fmt.Sprintf("	%v\n", &ratio)
+	}
+	return out + "}"
 }
